// convex/_lib/validators/addressFormats.generated.ts
export const ADDRESS_FORMAT_DATA: Record<string, AddressFormatData> = {
// ...
  ZZ: {
    fmt: '%N%n%O%n%A%n%C',
    require: 'AC',
    upper: 'C',
    regionType: 'province',
    cityType: 'city',
    postalType: 'postal',
  },
};